Core/DB/Quests: Fixed some quests for "The Crucible of Carnage" (Progress 50%)
*Note*:

[FIXED]
The Crucible of Carnage: The Bloodeye Bruiser! (Hurp'Derp - 100%)
The Crucible of Carnage: The Deadly Dragonmaw! (Torg Drakeflayer - 100%)
The Crucible of Carnage: The Wayward Wildhammer! (Sully Kneecapper - 100%)

[TODO]
The Crucible of Carnage: Calder's Creation! (Cadaver Collage)
The Crucible of Carnage: The Earl of Evisceration! (Lord Geoffery Tulvan)
The Crucible of Carnage: The Twilight Terror! (Emberscar the Devourer)