Core/Mechanics: Passive tank talent "Vengeance" completely redesigned following blizz patch notes:
"Vengeance has been redesigned slightly. It no longer ramps up slowly at the beginning of a fight.
Instead, the first melee attack taken by the tank generates Vengeance equal to 33% of the damage taken by that attack. In addition, as it updates periodically during the fight, it's always set to at least 33% of the damage taken by the tank in the last 2 seconds.
It still climbs from that point at the rate it did previously, still decays when damage is not taken, and still cannot exceed a maximum based on the health and Stamina of the tank."

Signed-off-by: john2308 <gianni.elio@hotmail.it>